Added 4GB Upload Limit Support
Added in the ability to chunk the upload of data to keep it under the 4GB threshold... Removed a couple sections on the uninstall action sets so that it doesn't copy ALL the files back down to the machine to perform an uninstall. This will need to be evaluated on a case by case basis.